+++ examples/openpose/openposeBasicDemo.cpp
@@ -1,0 +1,79 @@
+// ------------------------- OpenPose Library Tutorial - Real Time Pose Estimation -------------------------
+// C++ std library dependencies
+#include <chrono> // `std::chrono::` functions and classes, e.g. std::chrono::milliseconds
+#include <thread> // std::this_thread
+// Other 3rdparty dependencies
+#include <gflags/gflags.h> // DEFINE_bool, DEFINE_int32, DEFINE_int64, DEFINE_uint64, DEFINE_double, DEFINE_string
+#include <glog/logging.h> // google::InitGoogleLogging
+
+// OpenPose dependencies
+#include <openpose/headers.hpp>
+
+// See all the available parameter options withe the `--help` flag. E.g. `./build/examples/openpose/openpose.bin --help`.
+// Note: This command will show you flags for other unnecessary 3rdparty files. Check only the flags for the OpenPose
+// executable. E.g. for `openpose.bin`, look for `Flags from examples/openpose/openpose.cpp:`.
+// Producer
+DEFINE_int32(camera,                    -1,             "The camera index for cv::VideoCapture. Integer in the range [0, 9]. Select a negative"
+                                                        " number (by default), to auto-detect and open the first available camera.");
+DEFINE_string(camera_resolution,        "1280x720",     "Size of the camera frames to ask for.");
+DEFINE_double(camera_fps,               30.0,           "Frame rate for the webcam (only used when saving video from webcam). Set this value to the"
+                                                        " minimum value between the OpenPose displayed speed and the webcam real frame rate.");
+DEFINE_string(video,                    "",             "Use a video file instead of the camera. Use `examples/media/video.avi` for our default"
+                                                        " example video.");
+DEFINE_string(image_dir,                "",             "Process a directory of images. Use `examples/media/` for our default example folder with 20"
+                                                        " images. Read all standard formats (jpg, png, bmp, etc.).");
+// Display
+DEFINE_bool(no_gui_verbose,             false,          "Do not write text on output images on GUI (e.g. number of current frame and people). It"
+                                                        " does not affect the pose rendering.");
+DEFINE_bool(no_display,                 false,          "Do not open a display window. Useful if there is no X server and/or to slightly speed up"
+                                                        " the processing if visual output is not required.");
+
+int openPoseDemo()
+{
+    op::log("Starting pose estimation demo.", op::Priority::High);
+    const auto timerBegin = std::chrono::high_resolution_clock::now();
+
+    // Applying user defined configuration - Google flags to program variables
+    const auto producerSharedPtr = op::flagsToProducer(FLAGS_image_dir, FLAGS_video, FLAGS_camera, FLAGS_camera_resolution, FLAGS_camera_fps);
+    op::log("", op::Priority::Low, __LINE__, __FUNCTION__, __FILE__);
+
+    // OpenPose wrapper
+    op::log("Configuring OpenPose wrapper.", op::Priority::Low, __LINE__, __FUNCTION__, __FILE__);
+    op::Wrapper<std::vector<op::Datum>> opWrapper;
+    // Pose configuration (use WrapperStructPose{} for default and recommended configuration)
+    op::WrapperStructPose wrapperStructPose{};
+    wrapperStructPose.renderMode = op::RenderMode::Gpu;
+    // Producer (use default to disable any input)
+    const op::WrapperStructInput wrapperStructInput{producerSharedPtr};
+    // Consumer (comment or use default argument to disable any output)
+    const op::WrapperStructOutput wrapperStructOutput{!FLAGS_no_display, !FLAGS_no_gui_verbose};
+    // Configure wrapper
+    opWrapper.configure(wrapperStructPose, wrapperStructInput, wrapperStructOutput);
+    // Set to single-thread running (e.g. for debugging purposes)
+    // opWrapper.disableMultiThreading();
+
+    // Start processing
+    // Two different ways of running the program on multithread environment
+    op::log("Starting thread(s)", op::Priority::High);
+    opWrapper.exec();  // It blocks this thread until all threads have finished
+
+    // Measuring total time
+    const auto now = std::chrono::high_resolution_clock::now();
+    const auto totalTimeSec = (double)std::chrono::duration_cast<std::chrono::nanoseconds>(now-timerBegin).count() * 1e-9;
+    const auto message = "Real-time pose estimation demo successfully finished. Total time: " + std::to_string(totalTimeSec) + " seconds.";
+    op::log(message, op::Priority::High);
+
+    return 0;
+}
+
+int main(int argc, char *argv[])
+{
+    // Initializing google logging (Caffe uses it for logging)
+    google::InitGoogleLogging("openPoseDemo");
+
+    // Parsing command line flags
+    gflags::ParseCommandLineFlags(&argc, &argv, true);
+
+    // Running openPoseDemo
+    return openPoseDemo();
+}